feat(web): sidebar footer rework and analytics as a sibling view of the agents board
- Sidebar footer is now a stacked list: Analytics (admin), Settings (admin),
  and Sign out separated by a divider with destructive hover; the collapsed
  rail gets the same three as icon-only buttons. Agents list is flex-1 min-h-0
  so the footer stays pinned.
- New HomeShell shared by / and /analytics: same sidebar and hub WebSocket,
  right pane swaps between ConversationView and AnalyticsCommandCenter. View
  derives from the pathname; toggling uses history.pushState so nothing
  remounts, and the footer toggle labels the destination (Analytics/Agents)
  with an active state on analytics. Non-admin deep links bounce back to the
  agents view.
- AnalyticsCommandCenter drops the workspaceScope prop: a workspace Select
  (with All workspaces) is now the primary filter-bar control, backed by
  /api/workspaces and reflected in the URL query. Filter updates use
  history.replaceState instead of router.replace to keep the shell mounted.
- Analytics removed from Settings entirely: section, nav item, href special
  case, and render path; /analytics is a real static-export route rendering
  the shell, and workflow-runs-dialog links there now.
